requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=0.6,!=0.7,<1.0

Paste
PasteDeploy>=1.5.0
Routes>=1.12.3,!=2.0
anyjson>=0.3.3
argparse
Babel>=1.3
eventlet>=0.15.1
greenlet>=0.3.2
httplib2>=0.7.5
requests>=1.2.1,!=2.4.0
iso8601>=0.1.9
jsonrpclib
Jinja2
keystonemiddleware>=1.0.0
kombu>=2.4.8
netaddr>=0.7.12
python-neutronclient>=2.3.6,<3
SQLAlchemy>=0.8.4,<=0.8.99,>=0.9.7,<=0.9.99
WebOb>=1.2.3
python-keystoneclient>=0.10.0
alembic>=0.6.4
six>=1.7.0
stevedore>=0.14
oslo.config>=1.4.0.0a3
oslo.db>=0.4.0  # Apache-2.0
oslo.messaging>=1.4.0.0a3
oslo.rootwrap>=1.3.0.0a1

python-novaclient>=2.18.0

taskflow
